requirements.txt
matplotlib>=2.0.2
tensorflow>=1.2.1
glfw>=1.4.0
PyOpenGL>=3.1.0
pyassimp>=3.3